# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * DrainWatch civic safety platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type AuthSessionRole = typeof AuthSessionRole[keyof typeof AuthSessionRole];


export const AuthSessionRole = {
  citizen: 'citizen',
  officer: 'officer',
} as const;

export interface AuthSession {
  token: string;
  role: AuthSessionRole;
  name: string;
  email: string;
}

export type HazardType = typeof HazardType[keyof typeof HazardType];


export const HazardType = {
  waterlogging: 'waterlogging',
  blocked_drain: 'blocked_drain',
  open_manhole: 'open_manhole',
  sewage_overflow: 'sewage_overflow',
  damaged_road: 'damaged_road',
  other: 'other',
} as const;

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
  unknown: 'unknown',
} as const;

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  reported: 'reported',
  verified: 'verified',
  assigned: 'assigned',
  in_progress: 'in_progress',
  resolved: 'resolved',
  rejected: 'rejected',
} as const;

export interface RiskComponents {
  severity: number;
  recent_density: number;
  historical_frequency: number;
  rainfall: number;
  recency: number;
}

export type ReportRiskLevel = typeof ReportRiskLevel[keyof typeof ReportRiskLevel];


export const ReportRiskLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  critical: 'critical',
} as const;

export type ReportSource = typeof ReportSource[keyof typeof ReportSource];


export const ReportSource = {
  citizen: 'citizen',
  news: 'news',
  verified: 'verified',
} as const;

export interface Report {
  id: string;
  hazard_type: HazardType;
  severity: Severity;
  /**
     * @minimum 0
     * @maximum 100
     */
  risk_score: number;
  risk_level: ReportRiskLevel;
  risk_components: RiskComponents;
  status: ReportStatus;
  verification_status: string;
  latitude: number;
  longitude: number;
  location_name: string;
  description: string;
  source: ReportSource;
  is_demo: boolean;
  created_at: string;
  updated_at: string;
}

export interface ReportInput {
  hazard_type: HazardType;
  /** @minLength 10 */
  description: string;
  /** @minLength 2 */
  location_name: string;
  /**
     * @minimum 16
     * @maximum 18
     */
  latitude: number;
  /**
     * @minimum 77
     * @maximum 79
     */
  longitude: number;
}

export interface ReportStatusUpdate {
  status: ReportStatus;
}

export type PublicAlertSeverity = typeof PublicAlertSeverity[keyof typeof PublicAlertSeverity];


export const PublicAlertSeverity = {
  info: 'info',
  watch: 'watch',
  warning: 'warning',
} as const;

export interface PublicAlert {
  id: string;
  title: string;
  message: string;
  severity: PublicAlertSeverity;
  issued_at: string;
}

export interface DashboardStats {
  active_reports: number;
  critical_reports: number;
  verified_today: number;
  resolved_today: number;
  total_reports: number;
  rainfall_status: string;
  is_demo: boolean;
}

export interface ActivityItem {
  id: string;
  text: string;
  severity: Severity;
  status: ReportStatus;
  created_at: string;
}

export interface AuditLog {
  id: string;
  action: string;
  actor: string;
  report_id: string;
  created_at: string;
}

export type WeatherStatusStatus = typeof WeatherStatusStatus[keyof typeof WeatherStatusStatus];


export const WeatherStatusStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface WeatherStatus {
  status: WeatherStatusStatus;
  /** @nullable */
  rainfall_mm: number | null;
  /** @nullable */
  source: string | null;
  /** @nullable */
  observed_at: string | null;
}

export type ListOfficerReportsParams = {
status?: string;
severity?: Severity;
};