# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DrainWatch civic safety platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: public
    description: Public safety information
  - name: citizen
    description: Citizen portal operations
  - name: officer
    description: Authorized officer operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /public/alerts:
    get:
      operationId: listPublicAlerts
      tags: [public]
      summary: List public safety alerts
      responses:
        "200":
          description: Public alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicAlert"
  /public/reports:
    get:
      operationId: listPublicReports
      tags: [public]
      summary: List permitted public hazard reports
      responses:
        "200":
          description: Publicly visible reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
  /auth/citizen/login:
    post:
      operationId: citizenLogin
      tags: [citizen]
      summary: Sign in to the citizen portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Citizen session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [public]
      summary: Sign in and resolve the user's portal from their credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session with the server-validated portal role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/officer/login:
    post:
      operationId: officerLogin
      tags: [officer]
      summary: Sign in to the officer portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Officer session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /citizen/reports:
    get:
      operationId: listCitizenReports
      tags: [citizen]
      summary: List reports owned by the signed-in citizen
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Citizen reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "401":
          description: Citizen authentication required
    post:
      operationId: createCitizenReport
      tags: [citizen]
      summary: Submit a public hazard report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /citizen/reports/{id}:
    get:
      operationId: getCitizenReport
      tags: [citizen]
      summary: Get one owned citizen report
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Citizen report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Report not found
  /officer/dashboard:
    get:
      operationId: getOfficerDashboard
      tags: [officer]
      summary: Get officer dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Operational summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "403":
          description: Officer role required
  /officer/reports:
    get:
      operationId: listOfficerReports
      tags: [officer]
      summary: List authorized operational reports
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            $ref: "#/components/schemas/Severity"
      responses:
        "200":
          description: Operational reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "403":
          description: Officer role required
  /officer/reports/{id}:
    get:
      operationId: getOfficerReport
      tags: [officer]
      summary: Get full operational report details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Operational report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "403":
          description: Officer role required
        "404":
          description: Report not found
  /officer/reports/{id}/status:
    patch:
      operationId: updateOfficerReportStatus
      tags: [officer]
      summary: Update report workflow status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportStatusUpdate"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "403":
          description: Officer role required
  /officer/activity:
    get:
      operationId: listOfficerActivity
      tags: [officer]
      summary: List recent operational activity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /officer/audit-logs:
    get:
      operationId: listOfficerAuditLogs
      tags: [officer]
      summary: List officer audit events
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
  /officer/weather:
    get:
      operationId: getOfficerWeather
      tags: [officer]
      summary: Get rainfall context
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Weather data or explicit unavailable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherStatus"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session
  parameters:
    ReportId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required: [email, password]
    AuthSession:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [citizen, officer]
        name:
          type: string
        email:
          type: string
          format: email
      required: [token, role, name, email]
    HazardType:
      type: string
      enum:
        - waterlogging
        - blocked_drain
        - open_manhole
        - sewage_overflow
        - damaged_road
        - other
    Severity:
      type: string
      enum: [low, medium, high, critical, unknown]
    ReportStatus:
      type: string
      enum: [reported, verified, assigned, in_progress, resolved, rejected]
    RiskComponents:
      type: object
      properties:
        severity:
          type: number
        recent_density:
          type: number
        historical_frequency:
          type: number
        rainfall:
          type: number
        recency:
          type: number
      required: [severity, recent_density, historical_frequency, rainfall, recency]
    Report:
      type: object
      properties:
        id:
          type: string
        hazard_type:
          $ref: "#/components/schemas/HazardType"
        severity:
          $ref: "#/components/schemas/Severity"
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
        risk_level:
          type: string
          enum: [low, moderate, high, critical]
        risk_components:
          $ref: "#/components/schemas/RiskComponents"
        status:
          $ref: "#/components/schemas/ReportStatus"
        verification_status:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        location_name:
          type: string
        description:
          type: string
        source:
          type: string
          enum: [citizen, news, verified]
        is_demo:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - hazard_type
        - severity
        - risk_score
        - risk_level
        - risk_components
        - status
        - verification_status
        - latitude
        - longitude
        - location_name
        - description
        - source
        - is_demo
        - created_at
        - updated_at
    ReportInput:
      type: object
      properties:
        hazard_type:
          $ref: "#/components/schemas/HazardType"
        description:
          type: string
          minLength: 10
        location_name:
          type: string
          minLength: 2
        latitude:
          type: number
          minimum: 16
          maximum: 18
        longitude:
          type: number
          minimum: 77
          maximum: 79
      required: [hazard_type, description, location_name, latitude, longitude]
    ReportStatusUpdate:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ReportStatus"
      required: [status]
    PublicAlert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, watch, warning]
        issued_at:
          type: string
          format: date-time
      required: [id, title, message, severity, issued_at]
    DashboardStats:
      type: object
      properties:
        active_reports:
          type: integer
        critical_reports:
          type: integer
        verified_today:
          type: integer
        resolved_today:
          type: integer
        total_reports:
          type: integer
        rainfall_status:
          type: string
        is_demo:
          type: boolean
      required:
        - active_reports
        - critical_reports
        - verified_today
        - resolved_today
        - total_reports
        - rainfall_status
        - is_demo
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        status:
          $ref: "#/components/schemas/ReportStatus"
        created_at:
          type: string
          format: date-time
      required: [id, text, severity, status, created_at]
    AuditLog:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        actor:
          type: string
        report_id:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, action, actor, report_id, created_at]
    WeatherStatus:
      type: object
      properties:
        status:
          type: string
          enum: [available, unavailable]
        rainfall_mm:
          type: ["number", "null"]
        source:
          type: ["string", "null"]
        observed_at:
          type: ["string", "null"]
      required: [status, rainfall_mm, source, observed_at]