# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * DrainWatch civic safety platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatus } from './reportStatus';
import type { Severity } from './severity';

export interface ActivityItem {
  id: string;
  text: string;
  severity: Severity;
  status: ReportStatus;
  created_at: Date;
}