# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * DrainWatch civic safety platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: string;
  action: string;
  actor: string;
  report_id: string;
  created_at: Date;
}