# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * DrainWatch civic safety platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthSessionRole } from './authSessionRole';

export interface AuthSession {
  token: string;
  role: AuthSessionRole;
  name: string;
  email: string;
}