# lib/api-zod/src/generated/types/hazardType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * DrainWatch civic safety platform API
 * OpenAPI spec version: 0.1.0
 */

export type HazardType = typeof HazardType[keyof typeof HazardType];


export const HazardType = {
  waterlogging: 'waterlogging',
  blocked_drain: 'blocked_drain',
  open_manhole: 'open_manhole',
  sewage_overflow: 'sewage_overflow',
  damaged_road: 'damaged_road',
  other: 'other',
} as const;